requirements.txt
annotated-types==0.7.0 
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31    
cffi==1.17.1
click==8.1.8
colorama==0.4.6       
contourpy==1.3.1      
cryptography==44.0.2  
cycler==0.12.1        
dnspython==2.7.0      
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11     
fastapi-cli==0.0.7
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jedi==0.17.2
Jinja2==3.1.6
jwt==1.3.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
parso==0.7.1
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
